requirements.txt
couchbase
requests
pymongo
python-dateutil
faker